requirements.txt
streamlit
transformers
diffusers
huggingface_hub
torch
pdf2image
opencv-python-headless
Pillow
moviepy
python-dotenv
PyMuPDF
